backend/pyproject.toml
[tool.poetry]
name = "modularmind-rag"
version = "1.1.0"
description = "Modular Retrieval-Augmented Generation API"
authors = ["ModularMind Team <info@modularmind.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.20"
asyncpg = "^0.28.0"
alembic = "^1.12.0"
redis = "^5.0.0"
httpx = "^0.24.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
openai = "^1.1.1"
qdrant-client = "^1.6.0"
sentence-transformers = "^2.2.2"
faiss-cpu = "^1.7.4"
langchain = "^0.0.300"
tiktoken = "^0.5.1"
tenacity = "^8.2.3"
prometheus-client = "^0.17.1"
PyYAML = "^6.0.1"
pandas = "^2.1.0"
numpy = "^1.25.2"
Pillow = "^10.0.1"
pdf2image = "^1.16.3"
pytesseract = "^0.3.10"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
python-docx = "^1.0.0"
pymupdf = "^1.23.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
pytest-asyncio = "^0.21.1"
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.5.1"
ruff = "^0.0.290"
pre-commit = "^3.4.0"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
select = ["E", "F", "B", "I001"]
ignore = ["E501"]
line-length = 88
target-version = "py310"